refactor(kafka): split kafka package to run tests faster
