+++ internal/sdkprovider/service/kafkaschema/kafka_schema.go
@@ -1,4 +1,4 @@
-package kafka
+package kafkaschema
 
 import (
 	"context"
