+++ internal/sdkprovider/service/kafkaschema/kafka_schema_configuration.go
@@ -1,4 +1,4 @@
-package kafka
+package kafkaschema
 
 import (
 	"context"
