+++ internal/sdkprovider/service/kafkaschema/kafka_schema_configuration_data_source.go
@@ -1,4 +1,4 @@
-package kafka
+package kafkaschema
 
 import (
 	"context"
