+++ internal/sdkprovider/service/kafkatopic/kafka_topic.go
@@ -1,4 +1,4 @@
-package kafka
+package kafkatopic
 
 import (
 	"context"
