+++ internal/sdkprovider/service/kafkatopic/kafka_topic_cache_test.go
@@ -1,4 +1,4 @@
-package kafka
+package kafkatopic
 
 import (
 	"reflect"
