+++ internal/sdkprovider/service/kafkatopic/kafka_topic_data_source.go
@@ -1,4 +1,4 @@
-package kafka
+package kafkatopic
 
 import (
 	"context"
