+++ internal/sdkprovider/service/kafkatopic/kafka_topic_data_source_test.go
@@ -1,4 +1,4 @@
-package kafka_test
+package kafkatopic_test
 
 import (
 	"fmt"
